+++ Sources/MapboxMaps/Camera/CameraAnimationsManagerImpl.swift
@@ -133,7 +133,7 @@ internal final class CameraAnimationsManagerImpl: CameraAnimationsManagerProtoco
         return animator
     }
 
-    /// This function will handle the natural decelration of a gesture when there is a velocity provided. A use case for this is the pan gesture.
+    /// This function will handle the natural deceleration of a gesture when there is a velocity provided. A use case for this is the pan gesture.
     /// - Parameters:
     ///   - location: The initial location. This location will be simulated based on velocity and decelerationFactor.
     ///   - velocity: The initial velocity.