+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Mapbox welcomes participation and contributions from everyone.
 
 * Add easing curve parameter to `CameraAnimationsManager.fly(to:duration:curve:completion)`, make `TimingCurve` public with few more options.
 * Add `customData` field in Annotaion and deprecate `userInfo`. `userInfo` behaviour rolled back to v10 behaviour.
+* Fix MapView flickering during resizing.
+* Fix glitch in chained camera animations.
 
 ## 11.1.0-rc.1 - 04 January, 2024